Updating SlideMap
