+++ src/panels/stats-panel/stats-panel-container.css
@@ -18,6 +18,7 @@
   font-style: normal;
   z-index: 5;
   pointer-events: none;
+  font-size: 0.7rem;
 }
 
 .right-panel > * {
